Fixed IcePHP assert that would occur with a PHP debug executable when destroying multiple times the same communicator (ICE-8635)

+++ php/src/php5/Communicator.cpp
@@ -270,42 +270,36 @@ ZEND_METHOD(Ice_Communicator, destroy)
     CommunicatorInfoIPtr _this = Wrapper<CommunicatorInfoIPtr>::value(getThis() TSRMLS_CC);
     assert(_this);
 
-    //
-    // Remove all registrations.
-    //
-    {
-        IceUtilInternal::MutexPtrLock<IceUtil::Mutex> lock(_registeredCommunicatorsMutex);
-        for(vector<string>::iterator p = _this->ac->ids.begin(); p != _this->ac->ids.end(); ++p)
-        {
-            _registeredCommunicators.erase(*p);
-        }
-        _this->ac->ids.clear();
-    }
-
-    //
-    // We need to destroy any object|value factories installed by this request.
-    //
-    _this->destroyFactories(TSRMLS_C);
-
     Ice::CommunicatorPtr c = _this->getCommunicator();
     assert(c);
     CommunicatorMap* m = reinterpret_cast<CommunicatorMap*>(ICE_G(communicatorMap));
     assert(m);
-    assert(m->find(c) != m->end());
-    m->erase(c);
-
-    ValueFactoryManagerPtr vfm = ValueFactoryManagerPtr::dynamicCast(c->getValueFactoryManager());
-    assert(vfm);
-    vfm->destroy();
-
-    try
+    if(m->find(c) != m->end()) // If not already destroyed
     {
+        m->erase(c);
+
+        //
+        // Remove all registrations.
+        //
+        {
+            IceUtilInternal::MutexPtrLock<IceUtil::Mutex> lock(_registeredCommunicatorsMutex);
+            for(vector<string>::iterator p = _this->ac->ids.begin(); p != _this->ac->ids.end(); ++p)
+            {
+                _registeredCommunicators.erase(*p);
+            }
+            _this->ac->ids.clear();
+        }
+
+        //
+        // We need to destroy any object|value factories installed by this request.
+        //
+        _this->destroyFactories(TSRMLS_C);
+
+        ValueFactoryManagerPtr vfm = ValueFactoryManagerPtr::dynamicCast(c->getValueFactoryManager());
+        assert(vfm);
+        vfm->destroy();
+
         c->destroy();
-    }
-    catch(const IceUtil::Exception& ex)
-    {
-        throwException(ex TSRMLS_CC);
-        RETURN_NULL();
     }
 }
